+++ CHANGELOG.md
@@ -28,6 +28,8 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2026-09-10
+ * Improved Smart Stretch Tool by adding "bloc movements" option
 ### 2026-09-09
  * Added Smart Draw Front Panel Tool **(DEV team only)**
  * Updated PackingSolver dependency